+++ src/store.ts
@@ -26,6 +26,7 @@ const useStore = create<AppState>(
       duplicates: false,
       games: INITIAL_STATE.games,
       gamesList: GAMES,
+      missing: false,
       newVersion: '1',
       nicknames: false,
       rules: INITIAL_STATE.rules,
@@ -142,6 +143,10 @@ const useStore = create<AppState>(
         set((state) => {
           state.badges[state.selectedGame?.value][i].levelCap = newBadge;
         }),
+      editRule: (newRule: string, i: number) =>
+        set((state) => {
+          state.rules[state.selectedRuleset][i].content = newRule;
+        }),
       importState: (newAppState: Partial<AppState>) =>
         set((state) => {
           state.games = newAppState.games;
@@ -153,7 +158,7 @@ const useStore = create<AppState>(
         }),
       removeNew: () =>
         set((state) => {
-          state.newVersion = '2.5.0';
+          state.newVersion = '2.6.0';
         }),
       reorderRule: (destinationId: number, rule: TRule, sourceId: number) =>
         set((state) => {
@@ -188,6 +193,11 @@ const useStore = create<AppState>(
         set((state) => {
           state.text = text;
         }),
+      toggleMissing: () => {
+        set((state) => {
+          state.missing = !state.missing;
+        });
+      },
       toggleMode: () => {
         set((state) => {
           state.darkMode = !state.darkMode;