+++ CHANGELOG.md
@@ -1,14 +1,42 @@
+## Version 2.6.0
+
+🚀 New features
+
+- Show only missing encounters by enabling the option in **Settings**
+- Edit already created rules
+- PWA implementation
+- Initial analytics implementation
+
+🐛 Bug Fix
+
+- Responsive fixes for header
+- Format fix for sitemap
+
+💅 Enhancement
+
+- Close windows by clicking on the overlay
+- Icons for status options
+
+🏠 Internal
+
+- Bumped Cypress to 8.1.0
+- Bumped Eslint to 7.32.0
+- Bumped @types/node to 16.4.10
+
 ## Version 2.5.1
 
 🐛 Bug Fix
+
 - Fix to form field connection to Netlify
 
 💅 Enhancement
+
 - Meta tag updates
 - Footer and header included in page
 - Changed location of **Github** link to footer
 
-📝Documentation
+📝 Documentation
+
 - Changelog
 - Contribution guide
 - Code of Conduct
@@ -17,49 +45,59 @@
 ## Version 2.5.0
 
 🚀 New features
+
 - Rules Tab - Create, edit and share custom rulesets for your nuzlocke or use some popular defaults
 - PokéStats Tab - At a glance view of all captured, fainted, in-team, and failed to captured pokémon. Click on the pokémon to see more details!
 - New Team option in Status of pokémon. This links to the new Pokéstats Tab!
 
 🐛 Bug Fix
+
 - Duplicate pokémon alert no longer shows up on failed encounters
 
 💅 Enhancement
+
 - Share option is now a button inside the tracker
 
 ## Version 2.4.0
 
 🐛 Bug Fix
+
 - Sword and Shield edit level caps
 - Custom game Reset all fix
 
 💅 Enhancement
+
 - More options to better report a bug or suggest a feature
 - Disabled inputs and placeholder if no game is selected
 
 ## Version 2.3.0
 
 🚀 New features
+
 - Nickname option for encounters - can be found in Settings
 - Edit level caps for base games - from the pencil next to the game select
 - Report a bug or suggestion option!
 
 🐛 Bug Fix
+
 - Yellow level cap adjustments
 - Several bug fixes related to eliminating encounter locations
 
 ## Version 2.2.0
 
 🚀 New features
+
 - New share option for copy and pasting
 - New settings option to enable duplicate clause - alerts on dupes!
 - New Team option in Status of pokémon. This links to the new Pokéstats Tab!
 
 🐛 Bug Fix
+
 - SWSH encounters/level caps bug fix - if it still does not appear please delete
-site cache from your browser (IMPORTANT this will delete your other encounters
-saved on the site)
+  site cache from your browser (IMPORTANT this will delete your other encounters
+  saved on the site)
 
 💅 Enhancement
+
 - BW2 Easy/Normal/Challenge mode level caps - separated by slashes
 - HGSS and GSC Level caps up till Red